added functionality to the refresh button
