+++ src/views/index.ejs
@@ -44,7 +44,7 @@
               <th scope="col">URL</th>
             </tr>
           </thead>
-          <tbody>
+          <tbody id="results">
             <% show.forEach(item=> { %>
               <tr>
                 <th scope="row">
