+++ src/views/partials/head.ejs
@@ -11,6 +11,7 @@
     <link rel="stylesheet" type="text/css" href="/assets/css/fontawesome-all.min.css">
     <link rel="stylesheet" type="text/css" href="/assets/css/cookieconsent.min.css" />
     <link rel="stylesheet" type="text/css" href="/style.css">
+    <script src="/assets/js/main.js"></script>
     <script src="/assets/js/external/jquery-3.3.1.min.js"></script>
     <script src="/assets/js/external/bootstrap.bundle.min.js"></script>
 </head>
