+++ src/views/partials/header.ejs
@@ -7,9 +7,7 @@
         </button>
         <div class="text-right">
           <div class="collapse navbar-collapse" id="navbarCollapse">
-            <a href="/">
-              <button class="btn-primary" type="button">Refresh</button>
-            </a>
+              <button class="btn-primary" id="scrape" type="button">Refresh</button>
           </div>
         </div>
       </div>
